reorganize code for testing purposes
